requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-rest-swagger==2.2.0
django-allauth==0.41.0
django-rest-auth
django-cors-headers
django-taggit